requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
coloredlogs==15.0.1
confection==0.1.5
cryptography==43.0.1
cymem==2.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
googletrans==3.0.0
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
llama-index-core==0.11.14
llama-parse==0.5.6
marisa-trie==1.2.0
markdown-it-py==3.0.0
marker-pdf==0.2.17
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
onnxruntime==1.19.2
openai==1.43.0
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pdftext==0.3.13
pillow==10.4.0
preshed==3.0.9
protobuf==5.28.2
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
Pygments==2.18.0
pypdf==4.3.1
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
rich==13.8.0
safetensors==0.4.5
scikit-learn==1.4.2
scipy==1.14.1
selenium==4.25.0
setuptools==74.1.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
surya-ocr==0.5.0
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
texify==0.1.10
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.1
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.13.1
